refactor: 型サイズ縮小と constexpr 化、センチネル整理
- 各 enum class に明示的な uint8_t 基底型を付与
- *_MAX/*_MIN マクロを std::numeric_limits<T>::max()/min() に置換
- ヘッダ定義のメンバ関数群に constexpr を波及（インライン化補助）
- Node のフィールドを 4B/1B アライメント別に並べ替え、heading_level/indent_level を int8_t に縮小
- ViewportManager::Zoom* の戻り値を float センチネル(0.0)から bool changed に変更し、
  effect::ApplyThemeChange::new_zoom を撤去（zoom_index を SSOT 化）
- TitleBar の per-button bool hovered を撤去し、TitleBarHitZone で集約
- FileEntry の 3 つの bool を uint8_t flags_ ビットフラグに統合
- parser.cpp: blockquote_indices を parse_resource に載せて per-parse ヒープ確保を削減、
  monotonic 初期チャンクを 64KB→128KB に拡大
- Node::source_offset の "未設定" センチネルを kUnsetSourceOffset として共通化
- /simplify レビュー指摘対応: tautological テスト削除、requires-clause 整形、
  差分を narrate するだけのコメント削除

+++ src/layout/layout.cpp
@@ -98,7 +98,7 @@ float EstimateNodeHeight(const Node& node, const Theme& theme) noexcept
     const float line_height = theme.font_size_body * 1.5f;
     switch (node.type) {
     case NodeType::Heading: {
-        const int level = std::clamp(node.heading_level, 1, 6) - 1;
+        const int level = std::clamp(static_cast<int>(node.heading_level), 1, 6) - 1;
         return theme.font_size_h[level] * 1.5f;
     }
     case NodeType::CodeBlock: {
